+++ packages/echo-db/src/invitations/greeting-initiator.ts
@@ -27,7 +27,7 @@ import { greetingProtocolProvider } from './greeting-protocol-provider';
 import { GreetingState } from './greeting-responder';
 import { InvitationDescriptor, InvitationDescriptorType } from './invitation-descriptor';
 
-const log = debug('dxos:party-manager:greeting-initiator');
+const log = debug('dxos:echo:invitations:greeting-initiator');
 
 const DEFAULT_TIMEOUT = 30000;
 
@@ -78,12 +78,12 @@ export class GreetingInitiator {
     // Use the invitation ID as our peerId.
     // This is due to a bug in the protocol where the invitation id is omitted from the payload.
     // Therefore at present the greeter discovers the invitation id from session metadata, via the invitee's peer id.
-    // TODO(dboreham): invitation is actually invitationId.
+    // TODO(dboreham): invitation is actually invitationID.
     const localPeerId = invitation;
     log('Local PeerId:', keyToString(localPeerId));
     this._greeterPlugin = new GreetingCommandPlugin(localPeerId, (new Greeter()).createMessageHandler());
 
-    log('Connecting');
+    log(keyToString(localPeerId), 'connecting to', keyToString(swarmKey));
     const peerJoinedWaiter = waitForEvent(this._greeterPlugin, 'peer:joined',
       (remotePeerId: any) => remotePeerId && Buffer.from(responderPeerId).equals(remotePeerId), timeout);