+++ packages/model-factory/src/model.ts
@@ -18,8 +18,8 @@ export interface MutationWriteReceipt extends WriteReceipt {
  * Abstract base class for Models.
  * Models define a root message type, which is contained in the partent Item's message envelope.
  */
-export abstract class Model<T, U = void> {
-  protected readonly _modelUpdate = new Event<this>();
+export abstract class Model<T> {
+  public readonly modelUpdate = new Event<this>();
   private readonly _processor: NodeJS.WritableStream;
 
   private readonly _meta: ModelMeta;
@@ -66,7 +66,7 @@ export abstract class Model<T, U = void> {
   }
 
   subscribe (listener: (result: this) => void) {
-    return this._modelUpdate.on(listener);
+    return this.modelUpdate.on(listener);
   }
 
   /**
@@ -92,7 +92,7 @@ export abstract class Model<T, U = void> {
   async processMessage (meta: MutationMeta, message: T): Promise<void> {
     const modified = await this._processMessage(meta, message);
     if (modified) {
-      this._modelUpdate.emit(this);
+      this.modelUpdate.emit(this);
     }
   }
 
@@ -103,4 +103,13 @@ export abstract class Model<T, U = void> {
    * @param {Object} message
    */
   async abstract _processMessage (meta: MutationMeta, message: T): Promise<boolean>;
+
+  createSnapshot (): any {
+    throw new Error('This model does not support snapshots.');
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  async restoreFromSnapshot (snapshot: any): Promise<void> {
+    throw new Error('This model does not support snapshots');
+  }
 }